+++ src/main_win64.cpp
@@ -1,15 +1,15 @@
 /** OpenGL 3D Renderer
 
 TODO:
-    - orientation_to_direction and direction_to_orientation introduce a TON of error/uncertainty.
-    - BUG console command bug - commands get cut off when entered
+    - BUG console command bug - commands get cut off when entered - could be a memory bug?
     - Shadow mapping
         - directional shadows DONE
         - omnidirectional shadows for point lights and spot lights
         - multiple onnidirectional shadow maps
-        - multiple directional shadow maps
+        - CASCADED SHADOW MAPS https://developer.download.nvidia.com/SDK/10.5/opengl/src/cascaded_shadow_maps/doc/cascaded_shadow_maps.pdf
+            - make a test map for CSM. (e.g. field of trees or cubes all with shadows)
+        - option to completely disable shadows (test if we return to pre-shadows performance)
     - Skyboxes
-    - rewrite camera to use quaternion for rotation
     - Map Editor:
         - console command 'editor' to enter
         - quits the game inside the gamemode, and simply loads the map into the editor (keep camera in same transformation)
@@ -26,6 +26,7 @@ TODO:
       vertices in order to transform them on the screen (e.g. animate the text).
 
 Backlog:
+    - Implement GJK EPA collisions in a clone of this repo
     - option for some console messages to be displayed to game screen.
     - Memory management / custom memory allocator / replace all mallocs and callocs
     - Arrow rendering for debugging
@@ -138,19 +139,24 @@ texture_t g_font_atlas_c64;
 shader_lighting_t shader_common;
 
 shader_directional_shadow_map_t shader_directional_shadow_map;
-shader_base_t shader_debug_dir_shadow_map;
+shader_omni_shadow_map_t        shader_omni_shadow_map;
+shader_base_t                   shader_debug_dir_shadow_map;
 
 shader_orthographic_t shader_text;
 shader_orthographic_t shader_ui;
 shader_perspective_t shader_simple;
 material_t material_shiny = {4.f, 128.f };
 material_t material_dull = {0.5f, 1.f };
-light_point_t point_lights[2];
-light_spot_t spot_lights[2];
+// light_point_t point_lights[2];
+// light_spot_t spot_lights[2];
 
 unsigned int depthMap;
 unsigned int depthMapFBO;
 mat4 lightSpaceMatrix;
+
+unsigned int depthCubeMap;
+unsigned int depthCubeMapFBO;
+float depthCubeMapFarPlane = 25.f;
 
 static const char* vertex_shader_path = "shaders/default_phong.vert";
 static const char* frag_shader_path = "shaders/default_phong.frag";
@@ -467,9 +473,15 @@ internal void game_render()
         glActiveTexture(GL_TEXTURE1);
         glBindTexture(GL_TEXTURE_2D, depthMap);
 
+        glActiveTexture(GL_TEXTURE2);
+        glBindTexture(GL_TEXTURE_CUBE_MAP, depthCubeMap);
+
         glUniformMatrix4fv(shader_common.uid_directional_light_transform, 1, GL_FALSE, lightSpaceMatrix.ptr());
         glUniform1i(shader_common.uid_texture, 0);
         glUniform1i(shader_common.uid_directional_shadow_map, 1); // set directional shadow map to reference texture unit 1
+
+        glUniform1i(shader_common.uid_omni_shadow_map, 2);
+        glUniform1f(shader_common.uid_omni_far_plane, depthCubeMapFarPlane);
 
         gl_bind_point_lights(shader_common, loaded_map->pointlights.data(), (uint8)loaded_map->pointlights.size());
         gl_bind_spot_lights(shader_common, loaded_map->spotlights.data(), (uint8)loaded_map->spotlights.size());
@@ -621,6 +633,7 @@ int main(int argc, char* argv[]) // Our main entry point MUST be in this form wh
 
     gl_load_shader_program_from_file(shader_common, vertex_shader_path, frag_shader_path);
     gl_load_shader_program_from_file(shader_directional_shadow_map, "shaders/directional_shadow_map.vert", "shaders/directional_shadow_map.frag");
+    gl_load_shader_program_from_file(shader_omni_shadow_map, "shaders/omni_shadow_map.vert", "shaders/omni_shadow_map.geom", "shaders/omni_shadow_map.frag");
     gl_load_shader_program_from_file(shader_debug_dir_shadow_map, "shaders/debug_directional_shadow_map.vert", "shaders/debug_directional_shadow_map.frag");
     gl_load_shader_program_from_file(shader_text, text_vs_path, text_fs_path);
     gl_load_shader_program_from_file(shader_ui, ui_vs_path, ui_fs_path);
@@ -631,6 +644,27 @@ int main(int argc, char* argv[]) // Our main entry point MUST be in this form wh
     calculate_perspectivematrix(g_camera, 90.f);
     g_matrix_projection_ortho = projection_matrix_orthographic_2d(0.0f, (real32)g_buffer_width, (real32)g_buffer_height, 0.0f);
 
+    // point_lights[0].colour = { 0.0f, 1.0f, 0.0f };
+    // point_lights[0].position = { -4.f, 0.0f, 0.0f };
+    // point_lights[0].ambient_intensity = 0.f;
+    // point_lights[0].diffuse_intensity = 1.f;
+    // point_lights[1].colour = { 0.0f, 0.0f, 1.0f };
+    // point_lights[1].position = { 4.f, 0.0f, 0.0f };
+    // point_lights[1].ambient_intensity = 0.f;
+    // point_lights[1].diffuse_intensity = 1.f;
+    // debug_set_pointlights(point_lights, array_count(point_lights));
+    // spot_lights[0].position = { -4.f, 0.f, 0.f };
+    // spot_lights[0].ambient_intensity = 0.f;
+    // spot_lights[0].diffuse_intensity = 1.f;
+    // spot_lights[0].set_cutoff_in_degrees(45.f);
+    // spot_lights[0].orientation = direction_to_orientation(make_vec3(-1.f, -1.f, 0.f));
+    // spot_lights[1].position = { -2.f, 0.f, 0.f };
+    // spot_lights[1].ambient_intensity = 0.f;
+    // spot_lights[1].diffuse_intensity = 1.f;
+    // spot_lights[1].set_cutoff_in_degrees(45.f);
+    // spot_lights[1].orientation = direction_to_orientation(make_vec3(0.f, -1.f, 0.f));
+    // debug_set_spotlights(spot_lights, array_count(spot_lights));
+
     // TEMPORARY setting up maps/scenes (TODO replace once we are loading maps from disk)
     loaded_maps[0].directionallight.orientation = euler_to_quat(make_vec3(0.f, 30.f, -47.f) * KC_DEG2RAD);
     loaded_maps[0].directionallight.ambient_intensity = 0.3f;
@@ -640,8 +674,14 @@ int main(int argc, char* argv[]) // Our main entry point MUST be in this form wh
     loaded_maps[0].mainobject.pos = make_vec3(0.f, -6.f, 0.f);
     //loaded_maps[0].mainobject.orient = euler_to_quat(make_vec3(0.f, 30.f, -47.f)*KC_DEG2RAD);
     loaded_maps[0].mainobject.scale = make_vec3(0.04f, 0.04f, 0.04f);
-    loaded_maps[0].cam_start_pos = make_vec3(-47.44f, 66.29f, 9.65f);
-    loaded_maps[0].cam_start_rot = make_vec3(0.f, 30.f, -47.f);
+    loaded_maps[0].cam_start_pos = make_vec3(0.f, 0.f, 0.f);
+    loaded_maps[0].cam_start_rot = make_vec3(0.f, 211.f, -28.f);
+    light_point_t lm0pl;
+    lm0pl.colour = { 0.0f, 1.0f, 0.0f };
+    lm0pl.position = { -6.f, -2.0f, 4.0f };
+    lm0pl.ambient_intensity = 0.f;
+    lm0pl.diffuse_intensity = 1.f;
+    loaded_maps[0].pointlights.push_back(lm0pl);
 
     loaded_maps[1].directionallight.orientation = direction_to_orientation(make_vec3(2.f, -1.f, -2.f));
     loaded_maps[1].directionallight.ambient_intensity = 0.3f;
@@ -673,27 +713,6 @@ int main(int argc, char* argv[]) // Our main entry point MUST be in this form wh
 
     game_switch_map(0);
 
-    point_lights[0].colour = { 0.0f, 1.0f, 0.0f };
-    point_lights[0].position = { -4.f, 0.0f, 0.0f };
-    point_lights[0].ambient_intensity = 0.f;
-    point_lights[0].diffuse_intensity = 1.f;
-    point_lights[1].colour = { 0.0f, 0.0f, 1.0f };
-    point_lights[1].position = { 4.f, 0.0f, 0.0f };
-    point_lights[1].ambient_intensity = 0.f;
-    point_lights[1].diffuse_intensity = 1.f;
-    debug_set_pointlights(point_lights, array_count(point_lights));
-    spot_lights[0].position = { -4.f, 0.f, 0.f };
-    spot_lights[0].ambient_intensity = 0.f;
-    spot_lights[0].diffuse_intensity = 1.f;
-    spot_lights[0].set_cutoff_in_degrees(45.f);
-    spot_lights[0].orientation = direction_to_orientation(make_vec3(-1.f, -1.f, 0.f));
-    spot_lights[1].position = { -2.f, 0.f, 0.f };
-    spot_lights[1].ambient_intensity = 0.f;
-    spot_lights[1].diffuse_intensity = 1.f;
-    spot_lights[1].set_cutoff_in_degrees(45.f);
-    spot_lights[1].orientation = direction_to_orientation(make_vec3(0.f, -1.f, 0.f));
-    debug_set_spotlights(spot_lights, array_count(spot_lights));
-
 //////////// shadow implementation ////////////
     glGenFramebuffers(1, &depthMapFBO);
 
@@ -723,6 +742,50 @@ int main(int argc, char* argv[]) // Our main entry point MUST be in this form wh
     * view_matrix_look_at(make_vec3(-47.44f, 66.29f, 9.65f), make_vec3(-47.44f, 66.29f, 9.65f) + orientation_to_direction(loaded_maps[0].directionallight.orientation), make_vec3(0.f,1.f,0.f)); // TODO make up 0,0,1 if light is straight up or down
 
     temp_map_t* loaded_map = &loaded_maps[active_map_index];
+
+
+    glGenFramebuffers(1, &depthCubeMapFBO);
+
+    const unsigned int CUBE_SHADOW_WIDTH = 1024, CUBE_SHADOW_HEIGHT = 1024;
+
+    glGenTextures(1, &depthCubeMap);
+    glBindTexture(GL_TEXTURE_CUBE_MAP, depthCubeMap);
+    for (unsigned int i = 0; i < 6; ++i)
+    {
+        glTexImage2D(GL_TEXTURE_CUBE_MAP_POSITIVE_X + i, 0, GL_DEPTH_COMPONENT, 
+                     CUBE_SHADOW_WIDTH, CUBE_SHADOW_HEIGHT, 0, GL_DEPTH_COMPONENT, GL_FLOAT, NULL);
+    }
+    glTexParameteri(GL_TEXTURE_CUBE_MAP, GL_TEXTURE_MAG_FILTER, GL_NEAREST);
+    glTexParameteri(GL_TEXTURE_CUBE_MAP, GL_TEXTURE_MIN_FILTER, GL_NEAREST);
+    glTexParameteri(GL_TEXTURE_CUBE_MAP, GL_TEXTURE_WRAP_S, GL_CLAMP_TO_EDGE);
+    glTexParameteri(GL_TEXTURE_CUBE_MAP, GL_TEXTURE_WRAP_T, GL_CLAMP_TO_EDGE);
+    glTexParameteri(GL_TEXTURE_CUBE_MAP, GL_TEXTURE_WRAP_R, GL_CLAMP_TO_EDGE);
+
+    glBindFramebuffer(GL_FRAMEBUFFER, depthCubeMapFBO);
+    glFramebufferTexture(GL_FRAMEBUFFER, GL_DEPTH_ATTACHMENT, depthCubeMap, 0);
+    glDrawBuffer(GL_NONE);
+    glReadBuffer(GL_NONE);
+    glBindFramebuffer(GL_FRAMEBUFFER, 0);
+
+    float aspect = (float)CUBE_SHADOW_WIDTH/(float)CUBE_SHADOW_HEIGHT;
+    float nearPlane = 1.0f;
+    mat4 shadowProj = projection_matrix_perspective(90.f * KC_DEG2RAD, aspect, nearPlane, depthCubeMapFarPlane);
+
+    vec3 lightPos = loaded_maps[0].pointlights[0].position;
+    std::vector<mat4> shadowTransforms;
+    shadowTransforms.push_back(shadowProj * 
+                     view_matrix_look_at(lightPos, lightPos + WORLD_FORWARD_VECTOR, WORLD_DOWN_VECTOR));
+    shadowTransforms.push_back(shadowProj * 
+                     view_matrix_look_at(lightPos, lightPos + WORLD_BACKWARD_VECTOR, WORLD_DOWN_VECTOR));
+    shadowTransforms.push_back(shadowProj * 
+                     view_matrix_look_at(lightPos, lightPos + WORLD_UP_VECTOR, WORLD_RIGHT_VECTOR));
+    shadowTransforms.push_back(shadowProj * 
+                     view_matrix_look_at(lightPos, lightPos + WORLD_DOWN_VECTOR, WORLD_LEFT_VECTOR));
+    shadowTransforms.push_back(shadowProj * 
+                     view_matrix_look_at(lightPos, lightPos + WORLD_RIGHT_VECTOR, WORLD_DOWN_VECTOR));
+    shadowTransforms.push_back(shadowProj * 
+                     view_matrix_look_at(lightPos, lightPos + WORLD_LEFT_VECTOR, WORLD_DOWN_VECTOR));
+
 ////////////////////////////////////////////////////
 
     glEnable(GL_CULL_FACE);
@@ -758,6 +821,25 @@ int main(int argc, char* argv[]) // Our main entry point MUST be in this form wh
             render_meshgroup(loaded_map->mainobject.model);
 
             //glCullFace(GL_BACK);
+
+        glBindFramebuffer(GL_FRAMEBUFFER, 0);
+
+
+        gl_use_shader(shader_omni_shadow_map);
+        glViewport(0, 0, CUBE_SHADOW_WIDTH, CUBE_SHADOW_HEIGHT);
+        glBindFramebuffer(GL_FRAMEBUFFER, depthCubeMapFBO);
+            glClear(GL_DEPTH_BUFFER_BIT);
+            shader_omni_shadow_map.SetLightMatrices(shadowTransforms.data()); // sus if data works
+            shader_omni_shadow_map.SetLightPos(lightPos);
+            shader_omni_shadow_map.SetFarPlane(depthCubeMapFarPlane);
+
+            matrix_model = identity_mat4();
+            matrix_model = identity_mat4();
+            matrix_model *= translation_matrix(loaded_map->mainobject.pos);
+            matrix_model *= rotation_matrix(loaded_map->mainobject.orient);
+            matrix_model *= scale_matrix(loaded_map->mainobject.scale);
+            gl_bind_model_matrix(shader_omni_shadow_map, matrix_model.ptr());
+            render_meshgroup(loaded_map->mainobject.model);
 
         glBindFramebuffer(GL_FRAMEBUFFER, 0);