+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## Development Log
 
+### 2021-06-25
+- 
+
+### 2021-06-24
+- Omni-directional shadows for point lights and spot lights
+- Fix error in camera.cpp where FOV needed to be in radians not degrees.
+
 ### 2021-06-20
 - Directional shadows rendering implementation - needs refactoring and clean up
 - Counter-clockwise vertex winding order for triangles created in kc_truetypeassembler.h